Added!
- Added little check for decodeFile section.

+++ src/main/java/me/t3sl4/textfileencoderdemo/tfencoderdemo/MainController.java
@@ -295,13 +295,21 @@ public class MainController {
     public void decodeSelectedFile() {
         if(selectedDecFile != null && key != null && fileExtension.getText() != null) {
             try {
-                Path source = Paths.get(selectedDecFile.getAbsolutePath());
-                Path target = Paths.get(System.getProperty("user.home") + "/Desktop");
-                FileZIP.unzipFolder(source, target);
-                File delete = new File(selectedDecFile.getAbsolutePath());
-                delete.delete();
-                FileEncryption.decryptFile(newPath, key, fileExtension.getText());
-                selectedDecFilePath.setText(null);
+                if(selectedDecFile.getName().contains("encrypted")) {
+                    Path source = Paths.get(selectedDecFile.getAbsolutePath());
+                    Path target = Paths.get(System.getProperty("user.home") + "/Desktop");
+                    FileZIP.unzipFolder(source, target);
+                    File delete = new File(selectedDecFile.getAbsolutePath());
+                    delete.delete();
+                    FileEncryption.decryptFile(newPath, key, fileExtension.getText());
+                    selectedDecFilePath.setText(null);
+                } else {
+                    selectedDecFilePath.setText(null);
+                    alert.setTitle("HATA!");
+                    alert.setHeaderText("Şifreleme Algoritması Hatası.");
+                    alert.setContentText("Yalnızca şifrelenmiş dosyaları içeren arşivleri seçebilirsin.");
+                    alert.showAndWait();
+                }
             } catch (IOException e) {
                 e.printStackTrace();
             } catch (GeneralSecurityException e) {
